Re-raise candidate exceptions when raise_on_mismatches? is true
When writing a candidate codepath it's common to make silly
mistakes/typos that generate exceptions. In tests it would be nice to
see these exceptions printed straight to the console. Currently you only
get to see the exception buried inside a MismatchError message, like:

    ...lots of other text... @exception=#<ArgumentError: wrong number of arguments (0 for 1..2)>

There's no backtrace, which makes it hard to track down what went wrong.

Now, when raise_on_mismatches? is true, we'll re-raise any candidate
exceptions we can find. This lets you see the full exception and
backtrace in the test output.