Add pretty to_s formatting for MismatchError
Normally, overriding #to_s is a bad idea, but in this case these errors
are intended for a console or console-like test environment rather
than a live production system. When running tests, the most important
information is "why didn't this experiment work?", and to that end,
print out the values (inspected) or a backtrace of an exception, if one
was raised, when a MismatchError is printed to the screen.